refactor: extract useTaskFormHelpers hook
Moves editingTaskId/editingTaskText state, editingInputRef, and all
edit-task form handlers (startEditingTask, saveTaskTitle, cancelEditingTask,
handleEditKeyDown, applySuggestionForEdit, handleEditInputChange) out of
App.jsx into a dedicated useTaskFormHelpers hook. App.jsx reduced by ~280 lines.
